Move `runtime` into `service-data`
