+++ lib/service-data/runtime/get-runtime-data.unit.spec.js
@@ -3,14 +3,14 @@ const path = require('path')
 
 const {getRuntimeData} = require('./get-runtime-data')
 
-const metadata = path.resolve('data', 'runtime', 'metadata')
+const metadata = path.resolve('data/runtime/metadata')
 const source1 = {
   source: 'source1',
   path: metadata
 }
 
 const getTestData = name => {
-  const dataPath = path.resolve('data', `${name}.json`)
+  const dataPath = path.resolve(`data/${name}.json`)
   return require(dataPath)
 }
