+++ lib/service-data/runtime/inject-repeatable-pages.js
