+++ lib/service-data/runtime/merge-instances.js
