+++ lib/service-data/runtime/propagate-categories.js
