+++ lib/service-data/service-data.js
@@ -1,5 +1,3 @@
-require('@ministryofjustice/module-alias/register-module')(module)
-
 const path = require('path')
 const fs = require('fs')
 const jsonPath = require('jsonpath')
@@ -12,7 +10,7 @@ const replace = require('replace-in-file')
 
 const {FBError, deepClone} = require('@ministryofjustice/fb-utils-node')
 
-const {getRuntimeData} = require('~/fb-runner-node/runtime/get-runtime-data')
+const {getRuntimeData} = require('./runtime/get-runtime-data')
 
 class FBServiceDataError extends FBError {
   constructor (error = {}) {
